EP-1237 НОВОЕ ГОРЯЧЕЕ КОМПЛЕКСНОЕ ПИТАНИЕ

+++ processor/src/web/java/ru/axetta/ecafe/processor/web/partner/preorder/dataflow/ComplexListParam.java
@@ -11,15 +11,15 @@ import java.util.List;
  */
 public class ComplexListParam {
 
-    private Long idOfComplex;
+    private Integer idOfComplex;
     private Integer amount;
     private List<MenuItemParam> menuItems;
 
-    public Long getIdOfComplex() {
+    public Integer getIdOfComplex() {
         return idOfComplex;
     }
 
-    public void setIdOfComplex(Long idOfComplex) {
+    public void setIdOfComplex(Integer idOfComplex) {
         this.idOfComplex = idOfComplex;
     }
